feat: big commit (changed authority managing for users, general logic of products; added images, exception handlers and custom exceptions; fully updated api for products, tags, types, and category; added swagger api)

+++ src/main/java/com/diploma/qoldan/service/WishlistService.java
@@ -1,21 +1,16 @@
 package com.diploma.qoldan.service;
 
-import com.diploma.qoldan.dto.ProductDto;
 import com.diploma.qoldan.mapper.ProductMapper;
-import com.diploma.qoldan.model.Product;
-import com.diploma.qoldan.model.User;
-import com.diploma.qoldan.model.Wishlist;
+import com.diploma.qoldan.model.product.Product;
+import com.diploma.qoldan.model.user.User;
+import com.diploma.qoldan.model.wishlist.Wishlist;
 import com.diploma.qoldan.repository.ProductRepo;
 import com.diploma.qoldan.repository.UserRepo;
 import com.diploma.qoldan.repository.WishlistRepo;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -27,23 +22,27 @@ public class WishlistService {
 
     private final ProductMapper productMapper;
 
-    @Transactional
-    public List<ProductDto> getUserWishlist(String username) {
-        User user = userRepo.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        List<Wishlist> wishlist = repo.findAllByUser(user);
-        List<Product> productList = wishlist.stream().map(Wishlist::getProduct).toList();
-        return productList
-                .stream()
-                .map(productMapper::mapProductToDto)
-                .toList();
-    }
+//    @Transactional
+//    public List<ProductDto> getUserWishlist(String username) {
+//        User user = userRepo.findByEmail(username);
+//        if (user == null) {
+//            throw new UsernameNotFoundException("");
+//        }
+//        List<Wishlist> wishlist = repo.findAllByUser(user);
+//        List<Product> productList = wishlist.stream().map(Wishlist::getProduct).toList();
+//        return productList
+//                .stream()
+//                .map(productMapper::mapProductToDto)
+//                .toList();
+//    }
 
     @Transactional
     public void addProductToWishlist(String username, Long productId) throws Exception {
         Product product = productRepo.findById(productId);
-        User user = userRepo.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+        User user = userRepo.findByEmail(username);
+        if (user == null) {
+            throw new UsernameNotFoundException("");
+        }
 
         Wishlist wishlist = repo.findByUserAndProduct(user, product);
         if (wishlist != null)
@@ -58,8 +57,10 @@ public class WishlistService {
     @Transactional
     public void deleteProductFromWishlist(String username, Long productId) {
         Product product = productRepo.findById(productId);
-        User user = userRepo.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+        User user = userRepo.findByEmail(username);
+        if (user == null) {
+            throw new UsernameNotFoundException("");
+        }
         Wishlist wishlist = repo.findByUserAndProduct(user, product);
         repo.delete(wishlist);
     }
